+++ cmd/neofs-cli/modules/object/search.go
@@ -142,6 +142,12 @@ func parseSearchFilters(cmd *cobra.Command) (object.SearchFilters, error) {
 		fs.AddPhyFilter()
 	}
 
+	for _, f := range fs {
+		if f.Header() == object.FilterPayloadHomomorphicHash {
+			return nil, fmt.Errorf("%s filter target is prohibited starting from API 2.23", object.FilterPayloadHomomorphicHash)
+		}
+	}
+
 	return fs, nil
 }